create different components :)
