+++ ST/Hodakov Alexey/project/src/components/footer/copyrights.vue
@@ -1,0 +1,38 @@
+<template>
+  <div class="BgCopyrights">
+      <div class="container">
+        <div
+          class="copyrights d-flex py-4 justify-content-around flex-column align-items-center flex-sm-row justify-content-sm-between"
+        >
+          <div>&copy; 2020 Brand All Rights Reserved.</div>
+          <div class="copyrightsIcons pt-3 pt-sm-0">
+            <a class="facebookIcon">
+              <i class="fab fa-facebook-f"></i>
+            </a>
+            <a class="twitterIcon">
+              <i class="fab fa-twitter"></i>
+            </a>
+            <a class="linkedinIcon">
+              <i class="fab fa-linkedin-in"></i>
+            </a>
+            <a class="pinterestIcon">
+              <i class="fab fa-pinterest-p"></i>
+            </a>
+            <a class="googleIcon">
+              <i class="fab fa-google-plus-g"></i>
+            </a>
+          </div>
+        </div>
+      </div>
+    </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
