+++ ST/Hodakov Alexey/project/src/components/footer/footer.vue
@@ -1,0 +1,95 @@
+<template>
+    <footer class="container px-0">
+      <div class="row col-12 px-0 my-5 no-gutters">
+        <div class="col-12 px-0 col-lg-6">
+          <div class="logo d-flex justify-content-center pb-3 justify-content-lg-start">
+            <a href="#" class="pr-2">
+              <img src="img/logo.png" alt="logo" />
+            </a>
+            <a href="#" class="pr-3 pr-xl-3 pr-lg-0 pr-md-2 pr-sm-2 brand">
+              bran
+              <span>d</span>
+            </a>
+          </div>
+
+          <div class="d-flex flex-column flex-sm-row px-5 px-sm-0 flex-lg-column">
+            <div class="footerBrandText pb-4 pr-sm-3">
+              Objectively transition extensive data rather than cross functional solutions. Monotonectally
+              syndicate
+              multidisciplinary materials before go forward benefits. Intrinsicly syndicate an expanded array of
+              processes and cross-unit partnerships.
+            </div>
+            <div class="footerBrandText pb-4">
+              Efficiently plagiarize 24/365 action items and focused infomediaries.
+              Distinctively seize superior initiatives for wireless technologies. Dynamically optimize.
+            </div>
+          </div>
+        </div>
+        <div
+          class="bottomNav d-none justify-content-between flex-wrap col-lg-6 pl-lg-4 pt-3 justify-content-md-between pl-sm-0 d-sm-flex"
+        >
+          <div class="footerColumn">
+            <h2>COMPANY</h2>
+            <a href="#">Home</a>
+            <a href="#">Shop</a>
+            <a href="#">About</a>
+            <a href="#">How It Works</a>
+            <a href="#">Contact</a>
+          </div>
+          <div class="footerColumn">
+            <h2>INFORMATION</h2>
+            <a href="#">Tearms&amp;Condition</a>
+            <a href="#">Privacy Policy</a>
+            <a href="#">How to Buy</a>
+            <a href="#">How to Sell</a>
+            <a href="#">Promotion</a>
+          </div>
+          <div class="footerColumn pt-5 pt-sm-0">
+            <h2>SHOP CATEGORY</h2>
+            <a href="#">Men</a>
+            <a href="#">Women</a>
+            <a href="#">Child</a>
+            <a href="#">Apparel</a>
+            <a href="#">Browse All</a>
+          </div>
+        </div>
+
+        <div class="bottomNavXS col-12 d-flex flex-column d-sm-none">
+          <h2 class="d-flex justify-content-around">COMPANY</h2>
+          <div class="footerColumn d-flex justify-content-around px-5">
+            <a href="#">Home</a>
+            <a href="#">Shop</a>
+            <a href="#">About</a>
+            <a href="#">How It Works</a>
+            <a href="#">Contact</a>
+          </div>
+          <h2 class="d-flex justify-content-around pt-5">INFORMATION</h2>
+          <div class="footerColumn d-flex justify-content-around px-5">
+            <a href="#">Tearms&amp;Condition</a>
+            <a href="#">Privacy Policy</a>
+            <a href="#">How to Buy</a>
+            <a href="#">How to Sell</a>
+            <a href="#">Promotion</a>
+          </div>
+
+          <h2 class="d-flex justify-content-around pt-5">SHOP CATEGORY</h2>
+          <div class="footerColumn d-flex justify-content-around px-5 pt-sm-0">
+            <a href="#">Men</a>
+            <a href="#">Women</a>
+            <a href="#">Child</a>
+            <a href="#">Apparel</a>
+            <a href="#">Browse All</a>
+          </div>
+        </div>
+      </div>
+    </footer>
+</template>
+
+<script>
+export default {
+
+};
+</script>
+
+<style>
+</style>
