+++ ST/Hodakov Alexey/project/src/components/footer/subscribe.vue
@@ -1,0 +1,52 @@
+<template>
+  <div class="subscribeBg pt-5 pb-5 container-fluid">
+    <div class="container px-0">
+      <div class="row pb-5">
+        <div class="subscribeReview d-flex col-lg-6 px-0 mt-5 mb-5 pr-0 pr-lg-5">
+          <div class="col-2 col-md-2 d-flex align-self-start">
+            <img src="img/Avatar.png" alt="Avatar" />
+          </div>
+          <div class="col-10 pl-5 pl-md-2 pl-lg-5">
+            <blockquote class="reviewsTextMassage">
+              “Vestibulum quis porttitor dui! Quisque viverra nunc mi,
+              a pulvinar purus condimentum a. Aliquam condimentum mattis neque sed pretium”
+            </blockquote>
+            <div class="reviewsTextName mt-3">Bin Burhan</div>
+            <div class="reviewsTextFrom">Dhaka, Bd</div>
+            <div class="reviewsTextSwitch mt-5 px-0 col-5 col-xl-4 col-lg-5 col-md-3 col-sm-4">
+              <div></div>
+              <div></div>
+              <div></div>
+            </div>
+          </div>
+        </div>
+
+        <div
+          class="subscribeEmail d-flex justify-content-center flex-column align-items-center col-lg-6 border-left"
+        >
+          <div class="subscribeEmailHead">Subscribe</div>
+          <div class="subscribeEmailText pb-4">FOR OUR NEWLETTER AND PROMOTION</div>
+          <form class="subscribeForm input-group mb-3 pl-1 pl-sm-4">
+            <input
+              type="email"
+              class="form-control pt-4 pb-4 ml-xl-5 col-lg-7 offset-lg-0 ml-lg-4 col-md-6 offset-md-2 col-sm-9 offset-sm-0"
+              placeholder="Enter Your Email"
+              aria-label="Enter Your Email"
+              aria-describedby="basic-addon2"
+            />
+            <div class="input-group-append">
+              <button class="btn btn-outline-secondary" type="button">Subscribe</button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style>
+</style>
