+++ ST/Hodakov Alexey/project/src/components/header/navigation.vue
@@ -1,0 +1,297 @@
+<template>
+  <div class="nav container d-none d-md-flex justify-content-center mt-3">
+    <div class="navAll">
+      <a class="nav-link" href="#">Home</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">home</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">home</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">home</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">home</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">Man</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">Man</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Man</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">Man</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Man</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">woman</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">woman</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">woman</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">woman</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">woman</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">Kids</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">Kids</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Kids</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">Kids</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Kids</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">Accoseriese</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">Accoseriese</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Accoseriese</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">Accoseriese</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Accoseriese</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">Featured</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">Featured</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Featured</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">Featured</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Featured</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="navAll">
+      <a class="nav-link" href="#">Hot Deals</a>
+      <div class="dropNav">
+        <div class="dropNavIn">
+          <div class="dropNavItem">Hot Deals</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+          <a href="#" class="dropNavItem" type="button">Accessories</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Hot Deals</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href="#" class="dropNavItem" type="button">Blazers</a>
+          <div class="dropNavItem">Hot Deals</div>
+          <a href="#" class="dropNavItem" type="button">Denim</a>
+          <a href="#" class="dropNavItem" type="button">Leggings/Pants</a>
+          <a href="#" class="dropNavItem" type="button">Skirts/Shorts</a>
+        </div>
+        <div class="dropNavIn">
+          <div class="dropNavItem">Hot Deals</div>
+          <a href="#" class="dropNavItem" type="button">Dresses</a>
+          <a href="#" class="dropNavItem" type="button">Tops</a>
+          <a href="#" class="dropNavItem" type="button">Sweaters/Knits</a>
+          <a href="#" class="dropNavItem" type="button">Jackets/Coats</a>
+          <a href>
+            <div>Super sale!</div>
+            <img src="img/navDrop.jpg" alt="nav img" />
+          </a>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style>
+</style>
