+++ ST/Hodakov Alexey/project/src/components/header/newArrivals.vue
@@ -1,0 +1,24 @@
+<template>
+  <div class="newArrivalsBg pb-5 pt-5">
+      <div
+        class="px-md-0 container d-flex flex-column flex-sm-row justify-content-between align-items-center"
+      >
+        <div class="newArrivals">New Arrivals</div>
+        <div class="newArrivalsPath mt-4 mt-sm-0">
+          <a href="#">home</a>
+          <a href="#">men</a>
+          <a href="#">new arrivals</a>
+        </div>
+      </div>
+    </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
