+++ ST/Hodakov Alexey/project/src/layout/styles/styles.css
@@ -33,10 +33,7 @@
 .invisible {
   display: none !important;
 }
-#basket {
-  background: #fff !important;
-  padding: 10px;
-}
+
 #basket-toggler {
   width: 50px;
   height: 35px;
@@ -279,7 +276,6 @@
     width: 260px;
     left: -40px;
     z-index: 2;
-    border: 1px solid #f16d7f;
   }
   .fa-star,.fa-star-half-alt{
     color: #f16d7f;
